Refactor: Improve file naming and module structure
- Rename REAME.md to README.md (fix typo)
- Rename Tetris_module.py to game_logic.py (more descriptive)
- Rename Tetris_sub_module.py to helpers.py (clearer purpose)
- Update all import statements to reflect new module names
